Make corrections based on pylint and Richard's suggestions.
